# frontend/openapi/careeros.yaml
openapi: 3.0.3
info: {title: CareerOS API, version: 0.4.0}
servers: [{url: /}]
paths:
...
  /api/jobs/timeline:
    get:
      operationId: getJobTimeline
      tags: [Discovery]
      parameters: [{name: period,in: query,schema: {type: string,enum: [ALL,TODAY,YESTERDAY,LAST_7_DAYS,LAST_30_DAYS]}},{name: companyId,in: query,schema: {type: string,format: uuid}},{name: page,in: query,schema: {type: integer}},{name: size,in: query,schema: {type: integer}},{name: minimumScore,in: query,schema: {type: integer}},{name: remote,in: query,schema: {type: boolean}},{name: location,in: query,schema: {type: string}},{name: ats,in: query,schema: {type: string}},{name: sourceType,in: query,schema: {$ref: '#/components/schemas/JobSourceType'}}]
      responses: {'200': {description: Jobs, content: {application/json: {schema: {$ref: '#/components/schemas/JobPage'}}}}}
  /api/v1/jobs:
    get:
      operationId: listJobs
      tags: [Jobs]
      parameters: [{name: page,in: query,schema: {type: integer}},{name: size,in: query,schema: {type: integer}},{name: title,in: query,schema: {type: string}},{name: location,in: query,schema: {type: string}},{name: remote,in: query,schema: {type: boolean}},{name: sourceType,in: query,schema: {$ref: '#/components/schemas/JobSourceType'}}]
      responses: {'200': {description: Jobs, content: {application/json: {schema: {$ref: '#/components/schemas/JobPostingPage'}}}}}
  /api/v1/jobs/{id}:
    get: {operationId: getJob,tags: [Jobs],parameters: [{name: id,in: path,required: true,schema: {type: string,format: uuid}}],responses: {'200': {description: Job details,content: {application/json: {schema: {$ref: '#/components/schemas/JobPosting'}}}}}}
...
  /api/resumes/job/{jobId}:
    get: {operationId: getJobResumeHistory,tags: [Resumes],parameters: [{name: jobId,in: path,required: true,schema: {type: string,format: uuid}}],responses: {'200': {description: History,content: {application/json: {schema: {type: array,items: {$ref: '#/components/schemas/ResumeVersion'}}}}}}}
...
components:
  schemas:
    Page: {type: object,properties: {page: {type: integer},size: {type: integer},totalElements: {type: integer,format: int64},totalPages: {type: integer},last: {type: boolean}}}
    JobSourceType: {type: string,enum: [DIRECT_PROVIDER,AGGREGATOR]}
    JobSummary: {type: object,required: [id,title,company,remote,applyUrl,overallScore],properties: {id: {type: string,format: uuid},title: {type: string},company: {type: string},location: {type: string},remote: {type: boolean},postedDate: {type: string,format: date},applyUrl: {type: string,format: uri},sourceType: {$ref: '#/components/schemas/JobSourceType'},sourceUrl: {type: string,format: uri},overallScore: {type: integer},roleScore: {type: integer},technologyScore: {type: integer},locationScore: {type: integer},companyPriorityScore: {type: integer},recencyScore: {type: integer},matchExplanation: {type: array,items: {type: string}},applicationStatus: {$ref: '#/components/schemas/ApplicationStatus'}}}
    JobPosting: {type: object,required: [id,companyId,companyName,title,remote,applyUrl],properties: {id: {type: string,format: uuid},externalId: {type: string},companyId: {type: string,format: uuid},companyName: {type: string},title: {type: string},location: {type: string},employmentType: {type: string},remote: {type: boolean},salaryMin: {type: number},salaryMax: {type: number},salaryCurrency: {type: string},description: {type: string},postedDate: {type: string,format: date},applyUrl: {type: string},sourceType: {$ref: '#/components/schemas/JobSourceType'},sourceUrl: {type: string,format: uri},createdAt: {type: string,format: date-time}}}
    JobPage: {allOf: [{$ref: '#/components/schemas/Page'},{type: object,properties: {content: {type: array,items: {$ref: '#/components/schemas/JobSummary'}}}}]}
    JobPostingPage: {allOf: [{$ref: '#/components/schemas/Page'},{type: object,properties: {content: {type: array,items: {$ref: '#/components/schemas/JobPosting'}}}}]}
...
    ApplicationStatus: {type: string,enum: [WISHLIST,APPLIED,RECRUITER_CONTACTED,REFERRAL_REQUESTED,RECRUITER_SCREEN,TECHNICAL_INTERVIEW,SYSTEM_DESIGN,BEHAVIORAL,FINAL_INTERVIEW,OFFER,REJECTED,WITHDRAWN]}
...
    ResumeVersion: {type: object,required: [id,versionNumber,status,promptVersion,pdfGenerated,archived],properties: {id: {type: string,format: uuid},versionNumber: {type: integer},status: {type: string},jobId: {type: string,format: uuid},jobTitle: {type: string},applicationId: {type: string,format: uuid},modelUsed: {type: string},promptVersion: {type: string},generationDurationMs: {type: integer,format: int64},validationResult: {type: string},validationWarnings: {type: array,items: {type: string}},generatedBullets: {type: array,items: {type: string}},pdfGenerated: {type: boolean},archived: {type: boolean},createdAt: {type: string,format: date-time}}}
...