# src/main/java/com/careeros/discovery/web/DiscoveryController.java
package com.careeros.discovery.web;import com.careeros.discovery.application.DailyDiscoveryService;import com.careeros.common.web.PageResponse;import org.springframework.data.domain.*;import org.springframework.web.bind.annotation.*;import java.time.*;import java.util.*;
@RestController public class DiscoveryController{private final DailyDiscoveryService s;public DiscoveryController(DailyDiscoveryService s){this.s=s;}@GetMapping("/api/dashboard")public DailyDiscoveryService.Dashboard dashboard(){return s.dashboard();}@GetMapping("/api/reminders")public List<DailyDiscoveryService.Reminder>reminders(){return s.reminders();}@GetMapping("/api/analytics")public Map<String,Object>analytics(){return s.analytics();}@GetMapping("/api/jobs/timeline")public PageResponse<DailyDiscoveryService.JobSummary>timeline(@RequestParam(defaultValue="TODAY")Period period,@RequestParam(required=false)UUID companyId,@RequestParam(required=false)Integer minimumScore,@RequestParam(required=false)Boolean remote,@RequestParam(required=false)String location,@RequestParam(required=false)String ats,@RequestParam(required=false)com.careeros.job.domain.JobSourceType sourceType,Pageable pageable){LocalDate from=switch(period){case ALL->null;case TODAY->LocalDate.now();case YESTERDAY->LocalDate.now().minusDays(1);case LAST_7_DAYS->LocalDate.now().minusDays(7);case LAST_30_DAYS->LocalDate.now().minusDays(30);};return PageResponse.from(s.timeline(from,companyId,minimumScore,remote,location,ats,sourceType,pageable),x->x);}public enum Period{ALL,TODAY,YESTERDAY,LAST_7_DAYS,LAST_30_DAYS}}